# specs/openapi/data-service.yml
openapi: "3.1.0"
info:
  title: "Pociag Data Service API"
  version: "0.1.0"
  description: |
    Internal domain read service.
    Exposes curated schedules, operations, disruptions, and dictionaries.
    Not exposed to frontend directly; consumed by Gateway (BFF) and internal services.
servers:
  - url: http://localhost:8083
    description: Local development

# ── Tags ─────────────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/schedules/{routeId}:
    get:
      summary: Get route by internal ID
      description: Returns full route with stations, operating dates, and connections.
      operationId: getRouteById
      tags: [schedules]
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Full route details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/schedules/{routeId}/stations:
    get:
      summary: Get stations for a route
      description: Returns ordered list of stops for a given route.
      operationId: getRouteStations
      tags: [schedules]
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Ordered station list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteStationListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/disruptions:
    get:
      summary: Query disruptions
      description: Returns paginated disruptions matching filters.
      operationId: queryDisruptions
      tags: [disruptions]
      parameters:
        - name: dateFrom
          in: query
          description: Disruptions active on or after this date
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          description: Disruptions active on or before this date
          schema:
            type: string
            format: date
        - name: stationExternalIds
          in: query
          description: Filter by station IDs (start or end station matches)
          schema:
            type: string
        - name: typeCode
          in: query
          description: Filter by disruption type code
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated disruption list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisruptionListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/disruptions/{disruptionId}:
    get:
      summary: Get disruption by internal ID
      description: Returns full disruption with affected routes.
      operationId: getDisruptionById
      tags: [disruptions]
      parameters:
        - name: disruptionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Full disruption details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisruptionDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Dictionaries ───────────────────────────────────────────────────────────
  /api/v1/stations:
    get:
      summary: Query stations
      description: Returns stations with optional search and pagination.
      operationId: queryStations
      tags: [dictionaries]
      parameters:
        - name: search
          in: query
          description: Search by name prefix (trigram, case-insensitive)
          schema:
            type: string
        - name: city
          in: query
          description: Filter by city
          schema:
            type: string
        - name: externalIds
          in: query
          description: Comma-separated external IDs to fetch
          schema:
            type: string
        - name: hasCoordinates
          in: query
          description: When true, only return stations that have geographic coordinates
          schema:
            type: boolean
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Station list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StationListResponse"
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 1000
    Offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      additionalProperties: false
      properties:
        error:
          type: string
        message:
          type: string
        trace_id:
          type: string
    Pagination:
      type: object
      required: [total, limit, offset]
      additionalProperties: false
      properties:
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer

    # ── Dictionaries ───────────────────────────────────────────────────────────
    Station:
      type: object
      required: [id, external_id, name]
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        external_id:
          type: integer
        name:
          type: string
        city:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    StationListResponse:
      type: object
      required: [data, pagination]
      additionalProperties: false
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Station"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    RouteStation:
      type: object
      required: [station_external_id, order_number]
      additionalProperties: false
      properties:
        station_external_id:
          type: integer
        station_name:
          type: string
        order_number:
          type: integer
        arrival_time:
          type: string
          description: "ISO 8601 duration or HH:MM"
        departure_time:
          type: string
          description: "ISO 8601 duration or HH:MM"
        arrival_day:
          type: integer
        departure_day:
          type: integer
        platform:
          type: string
        track:
          type: string
        stop_type:
          type: string
        commercial_category:
          type: string
    RouteStationListResponse:
      type: object
      required: [data]
      additionalProperties: false
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RouteStation"
...
    RouteConnection:
      type: object
      required: [type_code, station_external_id]
      additionalProperties: false
      properties:
        external_connection_id:
          type: string
        type_code:
          type: string
        type_name:
          type: string
        station_external_id:
          type: integer
        station_name:
          type: string
        wagon_numbers:
          type: string
        train1_order_id:
          type: integer
        train2_order_id:
          type: integer
    RouteDetail:
      type: object
      required: [id, schedule_id, order_id, stations, operating_dates]
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        schedule_id:
          type: integer
        order_id:
          type: integer
        train_order_id:
          type: integer
        name:
          type: string
        carrier_code:
          type: string
        national_number:
          type: string
        international_arrival_num:
          type: string
        international_departure_num:
          type: string
        commercial_category_symbol:
          type: string
        stations:
          type: array
          items:
            $ref: "#/components/schemas/RouteStation"
        operating_dates:
          type: array
          items:
            type: string
            format: date
        connections:
          type: array
          items:
            $ref: "#/components/schemas/RouteConnection"

    # ── Operations ─────────────────────────────────────────────────────────────
...
    DisruptionSummary:
      type: object
      required: [id, external_disruption_id]
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        external_disruption_id:
          type: integer
          format: int64
        disruption_type_code:
          type: string
        disruption_type_name:
          type: string
        start_station_ext_id:
          type: integer
        start_station_name:
          type: string
        end_station_ext_id:
          type: integer
        end_station_name:
          type: string
        message:
          type: string
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        affected_routes_count:
          type: integer
    DisruptionListResponse:
      type: object
      required: [data, pagination]
      additionalProperties: false
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DisruptionSummary"
        pagination:
          $ref: "#/components/schemas/Pagination"
    DisruptionAffectedRoute:
      type: object
      required: [schedule_id, order_id]
      additionalProperties: false
      properties:
        schedule_id:
          type: integer
        order_id:
          type: integer
        train_order_id:
          type: integer
        operating_date:
          type: string
          format: date
        station_ext_id:
          type: integer
        station_name:
          type: string
        sequence_number:
          type: integer
    DisruptionDetail:
      type: object
      required: [id, external_disruption_id, affected_routes]
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        external_disruption_id:
          type: integer
          format: int64
        disruption_type_code:
          type: string
        disruption_type_name:
          type: string
        start_station_ext_id:
          type: integer
        start_station_name:
          type: string
        end_station_ext_id:
          type: integer
        end_station_name:
          type: string
        message:
          type: string
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        affected_routes:
          type: array
          items:
            $ref: "#/components/schemas/DisruptionAffectedRoute"